+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use AshAllenDesign\ShortURL\Facades\ShortURL;
-use Illuminate\Support\Facades\Route;
 
-Route::get('/'.ShortURL::prefix().'/{shortURLKey}', 'AshAllenDesign\ShortURL\Controllers\ShortURLController')->name('short-url.invoke');
+if (! config('short-url.disable_default_route')) {
+    ShortURL::routes();
+}