+++ src/Controllers/ShortURLController.php
@@ -6,16 +6,13 @@ use AshAllenDesign\ShortURL\Classes\Resolver;
 use AshAllenDesign\ShortURL\Models\ShortURL;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
-use Illuminate\Routing\Route;
 
 class ShortURLController
 {
     /**
-     * Redirect the user to the intended destination
-     * URL. If the default route has been disabled
-     * in the config but the controller has been
-     * reached using that route, return HTTP
-     * 404.
+     * Redirect the user to the intended destination URL. If the default
+     * route has been disabled in the config but the controller has
+     * been reached using that route, return HTTP 404.
      *
      * @param  Request  $request
      * @param  Resolver  $resolver
@@ -24,14 +21,6 @@ class ShortURLController
      */
     public function __invoke(Request $request, Resolver $resolver, string $shortURLKey): RedirectResponse
     {
-        /** @var Route $requestRoute */
-        $requestRoute = $request->route();
-
-        if ($requestRoute->getName() === 'short-url.invoke'
-            && config('short-url.disable_default_route')) {
-            abort(404);
-        }
-
         $shortURL = ShortURL::where('url_key', $shortURLKey)->firstOrFail();
 
         $resolver->handleVisit(request(), $shortURL);