+++ tests/Unit/Controllers/ShortURLControllerEmptyPrefixTest.php
@@ -5,21 +5,21 @@ namespace AshAllenDesign\ShortURL\Tests\Unit\Controllers;
 use AshAllenDesign\ShortURL\Models\ShortURL;
 use AshAllenDesign\ShortURL\Tests\Unit\TestCase;
 
-class ShortURLControllerPrefixTest extends TestCase
+class ShortURLControllerEmptyPrefixTest extends TestCase
 {
     protected function getEnvironmentSetUp($app)
     {
-        $app['config']->set('short-url.prefix', '/s');
+        $app['config']->set('short-url.prefix', null);
 
         parent::getEnvironmentSetUp($app);
     }
 
     /** @test */
-    public function visitor_is_redirected_to_the_destination_url_with_custom_prefix()
+    public function visitor_is_redirected_to_the_destination_url_with_null_prefix()
     {
         ShortURL::create([
             'destination_url'      => 'https://google.com',
-            'default_short_url'    => config('short-url.default_url').'/'.config('short-url.prefix').'/12345',
+            'default_short_url'    => config('short-url.default_url').'/12345',
             'url_key'              => '12345',
             'single_use'           => true,
             'track_visits'         => true,
@@ -27,6 +27,6 @@ class ShortURLControllerPrefixTest extends TestCase
             'activated_at'         => now()->subMinute(),
         ]);
 
-        $this->get('/s/12345')->assertStatus(301)->assertRedirect('https://google.com');
+        $this->get('/12345')->assertStatus(301)->assertRedirect('https://google.com');
     }
 }