+++ tests/Unit/Controllers/ShortURLControllerTest.php
@@ -7,7 +7,6 @@ use AshAllenDesign\ShortURL\Models\ShortURL;
 use AshAllenDesign\ShortURL\Models\ShortURLVisit;
 use AshAllenDesign\ShortURL\Tests\Unit\TestCase;
 use Illuminate\Foundation\Testing\RefreshDatabase;
-use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Event;
 
 class ShortURLControllerTest extends TestCase
@@ -25,7 +24,7 @@ class ShortURLControllerTest extends TestCase
     {
         ShortURL::create([
             'destination_url'      => 'https://google.com',
-            'default_short_url'    => config('app.url').'/short/12345',
+            'default_short_url'    => config('short-url.default_url').'/short/12345',
             'url_key'              => '12345',
             'single_use'           => true,
             'track_visits'         => true,
@@ -37,31 +36,13 @@ class ShortURLControllerTest extends TestCase
     }
 
     /** @test */
-    public function request_is_aborted_if_custom_routing_is_enabled_but_the_default_route_has_been_used()
-    {
-        Config::set('short-url.disable_default_route', true);
-
-        ShortURL::create([
-            'destination_url'      => 'https://google.com',
-            'default_short_url'    => config('app.url').'/short/12345',
-            'url_key'              => '12345',
-            'single_use'           => true,
-            'track_visits'         => true,
-            'redirect_status_code' => 301,
-            'activated_at'         => now()->subMinute(),
-        ]);
-
-        $this->get('/short/12345')->assertNotFound();
-    }
-
-    /** @test */
     public function event_is_dispatched_when_the_short_url_is_visited()
     {
         Event::fake();
 
         $shortURL = ShortURL::create([
             'destination_url'                => 'https://google.com',
-            'default_short_url'              => config('app.url').'/short/12345',
+            'default_short_url'              => config('short-url.default_url').'/short/12345',
             'url_key'                        => '12345',
             'single_use'                     => true,
             'forward_query_params'           => false,
@@ -101,7 +82,7 @@ class ShortURLControllerTest extends TestCase
     {
         ShortURL::create([
             'destination_url'      => 'https://google.com',
-            'default_short_url'    => config('app.url').'/short/12345',
+            'default_short_url'    => config('short-url.default_url').'/short/12345',
             'url_key'              => '12345',
             'single_use'           => true,
             'track_visits'         => true,
@@ -117,7 +98,7 @@ class ShortURLControllerTest extends TestCase
     {
         ShortURL::create([
             'destination_url'      => 'https://google.com',
-            'default_short_url'    => config('app.url').'/short/12345',
+            'default_short_url'    => config('short-url.default_url').'/short/12345',
             'url_key'              => '12345',
             'single_use'           => true,
             'track_visits'         => true,
@@ -134,7 +115,7 @@ class ShortURLControllerTest extends TestCase
     {
         ShortURL::create([
             'destination_url'      => 'https://google.com',
-            'default_short_url'    => config('app.url').'/short/12345',
+            'default_short_url'    => config('short-url.default_url').'/short/12345',
             'url_key'              => '12345',
             'single_use'           => true,
             'track_visits'         => true,
@@ -151,7 +132,7 @@ class ShortURLControllerTest extends TestCase
     {
         ShortURL::create([
             'destination_url'      => 'https://google.com',
-            'default_short_url'    => config('app.url').'/short/12345',
+            'default_short_url'    => config('short-url.default_url').'/short/12345',
             'url_key'              => '12345',
             'single_use'           => true,
             'track_visits'         => true,
@@ -168,7 +149,7 @@ class ShortURLControllerTest extends TestCase
     {
         ShortURL::create([
             'destination_url'      => 'https://google.com?param1=abc',
-            'default_short_url'    => config('app.url').'/short/12345',
+            'default_short_url'    => config('short-url.default_url').'/short/12345',
             'url_key'              => '12345',
             'forward_query_params' => false,
             'redirect_status_code' => 301,
@@ -181,6 +162,7 @@ class ShortURLControllerTest extends TestCase
 
     /**
      * @test
+     *
      * @dataProvider forwardQueryParamsProvider
      */
     public function visitor_is_redirected_to_the_destination_with_source_query_parameters_if_option_set_to_true(