+++ tests/Unit/Controllers/ShortURLDisableRouteTest.php
@@ -4,22 +4,25 @@ namespace AshAllenDesign\ShortURL\Tests\Unit\Controllers;
 
 use AshAllenDesign\ShortURL\Models\ShortURL;
 use AshAllenDesign\ShortURL\Tests\Unit\TestCase;
+use Illuminate\Foundation\Testing\LazilyRefreshDatabase;
 
-class ShortURLControllerPrefixTest extends TestCase
+class ShortURLDisableRouteTest extends TestCase
 {
+    use LazilyRefreshDatabase;
+
     protected function getEnvironmentSetUp($app)
     {
-        $app['config']->set('short-url.prefix', '/s');
+        $app['config']->set('short-url.disable_default_route', true);
 
         parent::getEnvironmentSetUp($app);
     }
 
     /** @test */
-    public function visitor_is_redirected_to_the_destination_url_with_custom_prefix()
+    public function request_is_aborted_if_custom_routing_is_enabled_but_the_default_route_has_been_used(): void
     {
         ShortURL::create([
             'destination_url'      => 'https://google.com',
-            'default_short_url'    => config('short-url.default_url').'/'.config('short-url.prefix').'/12345',
+            'default_short_url'    => config('short-url.default_url').'/short/12345',
             'url_key'              => '12345',
             'single_use'           => true,
             'track_visits'         => true,
@@ -27,6 +30,6 @@ class ShortURLControllerPrefixTest extends TestCase
             'activated_at'         => now()->subMinute(),
         ]);
 
-        $this->get('/s/12345')->assertStatus(301)->assertRedirect('https://google.com');
+        $this->get('/short/12345')->assertNotFound();
     }
 }